Moved organizations-related REST API tests into organizations.go
