+++ tests/rest/server.go
@@ -34,7 +34,6 @@ package tests
 import (
 	"encoding/json"
 	"fmt"
-	"reflect"
 	"strconv"
 	"strings"
 
@@ -145,33 +144,6 @@ func readOrganizationsFromResponse(f *frisby.Frisby) OrganizationsResponse {
 		}
 	}
 	return response
-}
-
-// checkOrganizationsEndpoint check if the end point to return list of organizations responds correctly to HTTP GET command
-func checkOrganizationsEndpoint() {
-	f := frisby.Create("Check the end point to return list of organizations by HTTP GET method").Get(apiURL + "organizations")
-	f.Send()
-	f.ExpectStatus(200)
-	f.ExpectHeader(contentTypeHeader, ContentTypeJSON)
-	organizationsResponse := readOrganizationsFromResponse(f)
-	if organizationsResponse.Status != "ok" {
-		f.AddError(fmt.Sprintf("Expected status is 'ok', but got '%s' instead", organizationsResponse.Status))
-	}
-	expectedOrglist := []int{1, 2, 3, 4}
-	if !reflect.DeepEqual(organizationsResponse.Organizations, expectedOrglist) {
-		f.AddError(fmt.Sprintf("Expected the following organizations %v, but got %v instead", expectedOrglist, organizationsResponse.Organizations))
-	}
-	f.PrintReport()
-}
-
-// checkOrganizationsEndpointWrongMethods check if the end point to return list of arganizations responds correctly to other methods than HTTP GET
-func checkOrganizationsEndpointWrongMethods() {
-	checkGetEndpointByOtherMethods(apiURL + "organizations")
-}
-
-func constructURLForOrganizationsClusters(organization int) string {
-	orgID := strconv.Itoa(organization)
-	return apiURL + "organizations/" + orgID + "/clusters"
 }
 
 // checkClustersEndpointForKnownOrganizations check if the end point to return list of clusters responds correctly to HTTP GET command
